+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Releases
 
+## 2.0.3
+  - Nueva versión! 🚀
+
 ## 2.0.2
   - Nueva versión! 🚀